se traslada las peticiones de comanda y login al archivo servicios
