+++ src/componentes/Comanda/Comanda.jsx
@@ -1,6 +1,8 @@
 import { useEffect, useState } from "react";
 import "../Comanda/comanda.css";
-import { showAlertOrderConfirm, showAlertError } from "../../alert/alerts";
+import { showAlertError } from "../../alert/alerts";
+import { peticionPostOrders } from "../../servicios/servicios";
+
 
 function Comanda({
   onMount,
@@ -39,8 +41,8 @@ function Comanda({
   };
 
   //guarda el pedido
-  function PeticionPostOrders() {
-    const bearerToken = localStorage.getItem("token");
+  function PostOrders() {
+    // const bearerToken = localStorage.getItem("token");
     if (!client) {
       showAlertError("Por favor ingrese el nombre del cliente");
       return;
@@ -49,36 +51,12 @@ function Comanda({
       showAlertError("Por favor seleccione un producto");
       return;
     }
-    const requestOptions = {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: "Bearer " + bearerToken,
-      },
-      body: JSON.stringify({
-        client: client,
-        products: order.productos,
-        status: "pending",
-        dataEntry: new Date().toLocaleString(),
-        cantidad: order.quantity,
-        startTime: new Date(),
-      }),
-    };
-
-    showAlertOrderConfirm().then((orderConfirmed) => {
-      if (orderConfirmed) {
-        fetch("http://localhost:8080/orders", requestOptions)
-          .then((response) => response.json())
-          .then((responseJson) => {
-            localStorage.removeItem("order");
-            onResetOrder();
-            setClient("");
-          })
-          .catch((error) => {
-            console.error(error);
-          });
-      }
-    });
+    peticionPostOrders(client, order)
+    .then((responseJson) => {
+      localStorage.removeItem("order");
+      onResetOrder();
+      setClient("");
+    })
   }
 
   return (
@@ -133,7 +111,7 @@ function Comanda({
           <div className="info-total">
             <p className="suma-precio">${totalCuenta}</p>
           </div>
-          <button className="boton-orden" onClick={PeticionPostOrders}>
+          <button className="boton-orden" onClick={PostOrders}>
             CREAR ORDEN
           </button>
         </div>
